// src/api/template/request/templateRequest.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface TemplateRequest {
  recordId: string;
  templateId: string;
  documentName: string;
  signers: Array<{
    order: number;
    email: string;
    role: string;
    nombre: string;
    subject?: string;
    message?: string;
    expiration_days?: number;
  }>;
  options?: {
    redirect_url?: string;
    redirect_decline_url?: string;
  };
} 